\name{print.FPDclustering}
\alias{print.FPDclustering}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print for FPDclustering objects
}
\description{
Lists the available components for the given object}
\usage{
\method{print}{FPDclustering}(x,...) 
}

\arguments{
  \item{x}{
an object of class FPDclustering
}
\item{...}{Additional parameters for the function ls}

}

\author{
Cristina Tortora
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

