\name{summary.FPDclustering}
\alias{summary.FPDclustering}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary for FPDclusteringt Objects
}
\description{
Number of elements per cluster.
}
\usage{
\method{summary}{FPDclustering}(object, ... ) 
}

\arguments{
  \item{object}{
an object of class FPDclustering
}
\item{...}{Additional parameters for the function paris}
 
}

\author{
Cristina Tortora
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

